refactor: LockedCharacter 컴포넌트 개별 파일로 분리
