+++ src/app/level/guide/Character.tsx
@@ -1,5 +1,5 @@
 import Image from 'next/image';
-import { css, cva, cx } from '@/styled-system/css';
+import { css, cx } from '@/styled-system/css';
 
 interface Props {
   width: number;
@@ -63,64 +63,6 @@ function Character(props: Props) {
 }
 
 export default Character;
-
-interface LockedCharacterProps {
-  level: number;
-  size: 'md' | 'lg';
-}
-
-export function LockedCharacter(props: LockedCharacterProps) {
-  return (
-    <div key={props.level} className={cx(imageWrapperCss)}>
-      <Image
-        src={`/assets/character-by-level/locked/${props.level}.svg`}
-        alt="locked"
-        width={180}
-        height={180}
-        className={cx(imageCss, lockCharacterImageCva({ size: props.size }))}
-      />
-      <Image
-        src={`/assets/character-by-level/locked/lock.svg`}
-        alt="locked"
-        width={180}
-        height={180}
-        className={cx(imageCss, lockCoverImageCva({ size: props.size }))}
-      />
-    </div>
-  );
-}
-
-const lockCharacterImageCva = cva({
-  base: { position: 'absolute', top: 0, bottom: 0, left: 0, right: 0, margin: 'auto', animation: 'fadeIn .7s' },
-  variants: {
-    size: {
-      lg: {
-        width: '202px',
-        height: '151px',
-      },
-      md: {
-        width: '76px',
-        height: '56.8px',
-      },
-    },
-  },
-});
-
-const lockCoverImageCva = cva({
-  base: {},
-  variants: {
-    size: {
-      md: {
-        width: '72px',
-        height: '72px',
-      },
-      lg: {
-        width: '180px',
-        height: '180px',
-      },
-    },
-  },
-});
 
 const imageWrapperCss = css({
   position: 'relative',
